requirements.txt
python-dotenv
requests
supabase
huggingface-hub
